+++ main.js
@@ -6557,7 +6557,6 @@ function renderThreadList() {
   this.inputEl = void 0;
   this.sendButtonEl = void 0;
   this.composerBarEl = void 0;
-  this.composerBarExpandEl = void 0;
   this.runSettings = void 0;
   this.toolBadgesEl = void 0;
   this.threadTitleEl = void 0;
@@ -8065,7 +8064,6 @@ var PiAgentView = class extends f4.ItemView {
     this.plugin = t;
     this.running = false;
     this.canceling = false;
-    this.composerBarExpanded = false;
     this.activityText = "Thinking";
     this.activityKind = "thinking";
     this.activityDetail = "";
@@ -8346,7 +8344,6 @@ var PiAgentView = class extends f4.ItemView {
     this.imageInputEl = void 0;
     this.sendButtonEl = void 0;
     this.composerBarEl = void 0;
-    this.composerBarExpandEl = void 0;
     this.runSettings = void 0;
     this.toolBadgesEl = void 0;
     this.threadTitleEl = void 0;
@@ -8647,22 +8644,8 @@ var PiAgentView = class extends f4.ItemView {
   updateComposerBarMode(e) {
     let t = this.composerBarEl;
     if (!t) return;
-    let n = e < 560,
-      s = e < 390;
-    if (!n && this.composerBarExpanded) this.composerBarExpanded = false;
-    t.toggleClass("is-compact", n);
-    t.toggleClass("is-narrow", s);
-    this.updateComposerBarExpansion();
-  }
-  updateComposerBarExpansion() {
-    let e = this.composerBarEl,
-      t = this.composerBarExpandEl;
-    if (!e || !t) return;
-    let n = this.composerBarExpanded && e.hasClass("is-compact");
-    e.toggleClass("is-expanded", n);
-    t.setAttr("aria-label", n ? "Collapse run options" : "Expand run options");
-    t.setAttr("title", n ? "Collapse run options" : "Expand run options");
-    (0, f4.setIcon)(t, n ? "chevrons-right" : "chevrons-left");
+    t.toggleClass("is-compact", e < 560);
+    t.toggleClass("is-narrow", e < 390);
   }
   renderImagePicker(parent) {
     const button = parent.createEl("button", {